Cleanup: separate service model and widgets in separate files
